edge_rsu/requirements.txt
# SmartV2X-CP Ultra — Edge RSU Dependencies
# ============================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
redis>=5.0.0
starlette>=0.27.0
httpx>=0.24.0
python-multipart>=0.0.6
